summarization-server/requirements.txt
fastapi==0.99.1
PyYAML==6.0.1
transformers==4.34.1
sentencepiece==0.1.99
aioredis==2.0.1
openai==1.7.2
pandas==1.3.5
python-multipart==0.0.6
unstructured==0.10.9
python-docx==0.8.11
pypdf==4.0.2
PyMUPDF==1.23.22
pdfminer.six==20231228
langchain==0.1.14
webvtt-py==0.4.6
faster-whisper==0.9.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
ffmpeg==1.4
llama-index-core==0.10.20
llama-index-readers-file==0.1.6
llama-index-vector-stores-faiss==0.1.2
llama-index-vector-stores-postgres==0.1.13
llama-index-embeddings-langchain==0.1.2
llama-index-retrievers-bm25==0.1.3
pydub==0.25.1
python-pptx==0.6.23
Werkzeug==3.0.0
docx2txt==0.8
rank-bm25==0.2.2
pymongo==4.2.0
okta_jwt==1.3.5
jwt==1.3.1
uvicorn[standard]

